Fix libmy and makefiles

+++ cs-392/src/my/my_strcmp.c
@@ -25,5 +25,5 @@ int my_strcmp(char* a, char* b){
 	
 	for( ; *a && *b && (*a == *b); a++, b++);
 	
-	return *a - *b;
+	return (int)(*a) - (int)(*b);
 }
